Split contacts and fragments tests
